# openapi.yaml
openapi: 3.1.0
info:
  title: YouTube Automation API
  description: |
    REST API for YouTube video automation system that manages video lifecycle from ideation to post-publication.

    This API provides comprehensive CRUD operations for video management across 8 distinct phases:
    - Phase 0: Published
    - Phase 1: Publish Pending
    - Phase 2: Edit Requested
    - Phase 3: Material Done
    - Phase 4: Started
    - Phase 5: Delayed
    - Phase 6: Sponsored Blocked
    - Phase 7: Ideas

    Each video progresses through phases with specific tasks and completion tracking.
  version: 1.0.0
  contact:
    name: Viktor Farcic
    url: https://github.com/vfarcic/youtube-automation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /api/videos:
    get:
      summary: List videos by phase
      description: Returns all videos for a given phase with computed progress data.
      operationId: getVideos
      tags:
        - Videos
      parameters:
        - name: phase
          in: query
          required: true
          schema:
            type: integer
          description: Phase number (0-7)
      responses:
        "200":
          description: List of enriched videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      summary: Create a video
      description: Creates a new video with the given name, category, and optional date.
      operationId: createVideo
      tags:
        - Videos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoRequest"
      responses:
        "201":
          description: Video created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/videos/{videoName}:
    get:
      summary: Get a video
      description: Returns a single video by name and category with computed progress.
      operationId: getVideo
      tags:
        - Videos
      parameters:
        - $ref: "#/components/parameters/VideoName"
        - $ref: "#/components/parameters/Category"
      responses:
        "200":
          description: Video details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update a video
      description: Replaces the full video data. Fields from the request body are merged over the existing video.
      operationId: updateVideo
      tags:
        - Videos
      parameters:
        - $ref: "#/components/parameters/VideoName"
        - $ref: "#/components/parameters/Category"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Partial or full video fields (JSON tags from storage.Video)
      responses:
        "200":
          description: Updated video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      summary: Patch a video aspect
      description: |
        Applies a partial update to a single aspect of a video. Only fields belonging
        to the specified aspect are accepted. Returns the full enriched video.
      operationId: patchVideoAspect
      tags:
        - Videos
      parameters:
        - $ref: "#/components/parameters/VideoName"
        - $ref: "#/components/parameters/Category"
        - name: aspect
          in: query
          required: true
          schema:
            type: string
            enum:
              - initial_details
              - work_progress
              - definition
              - post_production
              - publishing
              - post_publish
              - analysis
          description: Aspect key to patch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Field names (JSON tags) and their new values for the specified aspect
      responses:
        "200":
          description: Updated video with optional sync warning
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Delete a video
      description: Permanently deletes a video by name and category.
      operationId: deleteVideo
      tags:
        - Videos
      parameters:
        - $ref: "#/components/parameters/VideoName"
        - $ref: "#/components/parameters/Category"
      responses:
        "204":
          description: Video deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/drive/upload/video/{videoName}:
    post:
      summary: Upload video to Google Drive
      description: |
        Uploads a video file to Google Drive and updates the video's VideoFile and VideoDriveFileID fields.
        Creates a per-video subfolder if a root folder is configured.
      operationId: driveUploadVideo
      tags:
        - Drive
      parameters:
        - $ref: "#/components/parameters/VideoName"
        - $ref: "#/components/parameters/Category"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - video
              properties:
                video:
                  type: string
                  format: binary
                  description: Video file
      responses:
        "200":
          description: Upload successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  driveFileId:
                    type: string
                  videoFile:
                    type: string
                    example: "drive://abc123"
                  syncWarning:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "501":
          $ref: "#/components/responses/NotImplemented"
  /api/drive/download/video/{videoName}:
    get:
      summary: Download video from Google Drive
      description: Streams the video file from Google Drive. Returns the binary file with appropriate Content-Type and Content-Disposition headers.
      operationId: driveDownloadVideo
      tags:
        - Drive
      parameters:
        - $ref: "#/components/parameters/VideoName"
        - $ref: "#/components/parameters/Category"
      responses:
        "200":
          description: Video file stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "501":
          $ref: "#/components/responses/NotImplemented"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token authentication. Set the API_TOKEN environment variable on the server.
        If API_TOKEN is empty, authentication is disabled. The /health endpoint is always public.
  parameters:
    VideoName:
      name: videoName
      in: path
      required: true
      schema:
        type: string
      description: Video name (filename without extension)
    Category:
      name: category
      in: query
      required: true
      schema:
        type: string
      description: Video category
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized - missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotImplemented:
      description: Feature not configured (e.g., Google Drive or publishing service not set up)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        detail:
          type: string
          description: Additional error details
...
    ProgressInfo:
      type: object
      properties:
        completed:
          type: integer
        total:
          type: integer
    VideoResponse:
      type: object
      description: Full video data with computed phase and per-aspect progress
      properties:
        id:
          type: string
          description: "Composite ID: category/name"
          example: "devops/my-video"
        phase:
          type: integer
          description: Computed lifecycle phase (0-7)
        init:
          $ref: "#/components/schemas/ProgressInfo"
        work:
          $ref: "#/components/schemas/ProgressInfo"
        define:
          $ref: "#/components/schemas/ProgressInfo"
        edit:
          $ref: "#/components/schemas/ProgressInfo"
        publish:
          $ref: "#/components/schemas/ProgressInfo"
        postPublish:
          $ref: "#/components/schemas/ProgressInfo"
        syncWarning:
          type: string
          description: Warning message if git sync failed or is not configured
        pullWarning:
          type: string
          description: |
            Warning message if the throttled `git pull --rebase` triggered when
            opening the video detail page failed. The local copy is still
            returned; the field is omitted on success or when git sync is not
            configured.
      additionalProperties: true
...
    CreateVideoRequest:
      type: object
      required:
        - name
        - category
      properties:
        name:
          type: string
          description: Video name (used as filename)
        category:
          type: string
          description: Video category
        date:
          type: string
          description: Optional publish date (YYYY-MM-DD)
...